back-end/requirement.txt
accelerate
diffusers
fastapi
langchain
langchain-core
langchain-google-genai
langchain-openai
langgraph
numpy
Pillow
python-multipart
torch --index-url https://download.pytorch.org/whl/cu129
uvicorn